requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
